Porting to Cygwin, first attempt
Add Cygwin as a recognized platform.  Mostly copy the FreeBSD case.

This builds but fails to work properly because of an mmap problem on
Cygwin: You can't remap without first unmapping.  See

  https://cygwin.com/pipermail/cygwin/2018-February/236044.html

It's probably possible to work around this by replacing mmap and
friends by direct calls to VirtualAlloc and friends.